pyproject.toml
[project]
name = "langgraph-framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-openai>=0.3.27",
    "langgraph>=0.5.0",
    "openai>=1.93.0",
    "python-dotenv>=1.1.1",
]
